requirements.txt
Flask
yfinance
pandas
plotly
mysql-connector-python
textblob
nltk
requests
werkzeug
